requirements.txt
asgiref==3.8.1
Django==5.0.7
pillow==10.4.0
python-slugify==8.0.4
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.12.2
